pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "chemnlp"
description = "Open source chemistry dataset & LLM"
readme = "README.md"
requires-python = "==3.9.*"
dependencies = [
        "pandas",
        "pydantic",
]

dynamic = ["version"]

[project.optional-dependencies]
dev = [
        "black[jupyter]>=22.1.0",
        "fire",
        "flake8>=3.8.3",
        "isort>=5.0.0",
        "pre-commit",
        "pydantic_yaml<=0.11.2",
        "pytest",
        "pubchempy",
]

dataset_creation = [
        "PyTDC",
        "rdkit-pypi",
        "ruamel.yaml",
        "modin[all]",
        "selfies",
        "deepsmiles",
        "pubchempy",
        "bioc",
        "pylatexenc",
        "canonicalize_psmiles@git+https://github.com/Ramprasad-Group/canonicalize_psmiles.git",
        "rxn-chem-utils",
      #  "safe-mol",
        "backoff",
        "givemeconformer",
]

training = [
        "deepspeed>=0.8.2", # https://github.com/microsoft/DeepSpeed/pull/2863
        "nvidia-ml-py3",
        "protobuf<3.20",
        "s3fs",
        "boto3<=1.26.90", # https://github.com/boto/boto3/issues/3648
]

tokenisation = [
        "zstandard",
        "apache_beam",
        "mwparserfromhell",
        "jsonlines",
]

[tool.setuptools_scm]
version_scheme = "post-release"
